knowledge-base-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-docx==1.1.0
pypdf2==3.0.1
redis==5.0.1
celery==5.3.6
httpx==0.26.0
